requirements.txt
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
seaborn==0.13.2
